Optifine 13

+++ src/main/java/net/optifine/model/ModelSprite.java
@@ -37,10 +37,10 @@ public class ModelSprite
         this.sizeY = sizeY;
         this.sizeZ = sizeZ;
         this.sizeAdd = sizeAdd;
-        this.minU = textureOffsetX / modelRenderer.textureWidth;
-        this.minV = textureOffsetY / modelRenderer.textureHeight;
-        this.maxU = (textureOffsetX + sizeX) / modelRenderer.textureWidth;
-        this.maxV = (textureOffsetY + sizeY) / modelRenderer.textureHeight;
+        this.minU = (float)textureOffsetX / modelRenderer.textureWidth;
+        this.minV = (float)textureOffsetY / modelRenderer.textureHeight;
+        this.maxU = (float)(textureOffsetX + sizeX) / modelRenderer.textureWidth;
+        this.maxV = (float)(textureOffsetY + sizeY) / modelRenderer.textureHeight;
     }
 
     public void render(Tessellator tessellator, float scale)
@@ -63,7 +63,7 @@ public class ModelSprite
             f3 = this.minV;
         }
 
-        renderItemIn2D(tessellator, f, f2, f1, f3, this.sizeX, this.sizeY, scale * this.sizeZ, this.modelRenderer.textureWidth, this.modelRenderer.textureHeight);
+        renderItemIn2D(tessellator, f, f2, f1, f3, this.sizeX, this.sizeY, scale * (float)this.sizeZ, this.modelRenderer.textureWidth, this.modelRenderer.textureHeight);
         GlStateManager.translate(-this.posX * scale, -this.posY * scale, -this.posZ * scale);
     }
 
@@ -76,36 +76,36 @@ public class ModelSprite
 
         float f = maxU - minU;
         float f1 = maxV - minV;
-        double d0 = MathHelper.abs(f) * (texWidth / 16.0F);
-        double d1 = MathHelper.abs(f1) * (texHeight / 16.0F);
+        double d0 = (double)(MathHelper.abs(f) * (texWidth / 16.0F));
+        double d1 = (double)(MathHelper.abs(f1) * (texHeight / 16.0F));
         WorldRenderer worldrenderer = tess.getWorldRenderer();
         GL11.glNormal3f(0.0F, 0.0F, -1.0F);
         worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-        worldrenderer.pos(0.0D, d1, 0.0D).tex(minU, maxV).endVertex();
-        worldrenderer.pos(d0, d1, 0.0D).tex(maxU, maxV).endVertex();
-        worldrenderer.pos(d0, 0.0D, 0.0D).tex(maxU, minV).endVertex();
-        worldrenderer.pos(0.0D, 0.0D, 0.0D).tex(minU, minV).endVertex();
+        worldrenderer.pos(0.0D, d1, 0.0D).tex((double)minU, (double)maxV).endVertex();
+        worldrenderer.pos(d0, d1, 0.0D).tex((double)maxU, (double)maxV).endVertex();
+        worldrenderer.pos(d0, 0.0D, 0.0D).tex((double)maxU, (double)minV).endVertex();
+        worldrenderer.pos(0.0D, 0.0D, 0.0D).tex((double)minU, (double)minV).endVertex();
         tess.draw();
         GL11.glNormal3f(0.0F, 0.0F, 1.0F);
         worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-        worldrenderer.pos(0.0D, 0.0D, width).tex(minU, minV).endVertex();
-        worldrenderer.pos(d0, 0.0D, width).tex(maxU, minV).endVertex();
-        worldrenderer.pos(d0, d1, width).tex(maxU, maxV).endVertex();
-        worldrenderer.pos(0.0D, d1, width).tex(minU, maxV).endVertex();
+        worldrenderer.pos(0.0D, 0.0D, (double)width).tex((double)minU, (double)minV).endVertex();
+        worldrenderer.pos(d0, 0.0D, (double)width).tex((double)maxU, (double)minV).endVertex();
+        worldrenderer.pos(d0, d1, (double)width).tex((double)maxU, (double)maxV).endVertex();
+        worldrenderer.pos(0.0D, d1, (double)width).tex((double)minU, (double)maxV).endVertex();
         tess.draw();
-        float f2 = 0.5F * f / sizeX;
-        float f3 = 0.5F * f1 / sizeY;
+        float f2 = 0.5F * f / (float)sizeX;
+        float f3 = 0.5F * f1 / (float)sizeY;
         GL11.glNormal3f(-1.0F, 0.0F, 0.0F);
         worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
 
         for (int i = 0; i < sizeX; ++i)
         {
-            float f4 = (float)i / sizeX;
+            float f4 = (float)i / (float)sizeX;
             float f5 = minU + f * f4 + f2;
-            worldrenderer.pos(f4 * d0, d1, width).tex(f5, maxV).endVertex();
-            worldrenderer.pos(f4 * d0, d1, 0.0D).tex(f5, maxV).endVertex();
-            worldrenderer.pos(f4 * d0, 0.0D, 0.0D).tex(f5, minV).endVertex();
-            worldrenderer.pos(f4 * d0, 0.0D, width).tex(f5, minV).endVertex();
+            worldrenderer.pos((double)f4 * d0, d1, (double)width).tex((double)f5, (double)maxV).endVertex();
+            worldrenderer.pos((double)f4 * d0, d1, 0.0D).tex((double)f5, (double)maxV).endVertex();
+            worldrenderer.pos((double)f4 * d0, 0.0D, 0.0D).tex((double)f5, (double)minV).endVertex();
+            worldrenderer.pos((double)f4 * d0, 0.0D, (double)width).tex((double)f5, (double)minV).endVertex();
         }
 
         tess.draw();
@@ -114,13 +114,13 @@ public class ModelSprite
 
         for (int j = 0; j < sizeX; ++j)
         {
-            float f7 = (float)j / sizeX;
+            float f7 = (float)j / (float)sizeX;
             float f10 = minU + f * f7 + f2;
-            float f6 = f7 + 1.0F / sizeX;
-            worldrenderer.pos(f6 * d0, 0.0D, width).tex(f10, minV).endVertex();
-            worldrenderer.pos(f6 * d0, 0.0D, 0.0D).tex(f10, minV).endVertex();
-            worldrenderer.pos(f6 * d0, d1, 0.0D).tex(f10, maxV).endVertex();
-            worldrenderer.pos(f6 * d0, d1, width).tex(f10, maxV).endVertex();
+            float f6 = f7 + 1.0F / (float)sizeX;
+            worldrenderer.pos((double)f6 * d0, 0.0D, (double)width).tex((double)f10, (double)minV).endVertex();
+            worldrenderer.pos((double)f6 * d0, 0.0D, 0.0D).tex((double)f10, (double)minV).endVertex();
+            worldrenderer.pos((double)f6 * d0, d1, 0.0D).tex((double)f10, (double)maxV).endVertex();
+            worldrenderer.pos((double)f6 * d0, d1, (double)width).tex((double)f10, (double)maxV).endVertex();
         }
 
         tess.draw();
@@ -129,13 +129,13 @@ public class ModelSprite
 
         for (int k = 0; k < sizeY; ++k)
         {
-            float f8 = (float)k / sizeY;
+            float f8 = (float)k / (float)sizeY;
             float f11 = minV + f1 * f8 + f3;
-            float f13 = f8 + 1.0F / sizeY;
-            worldrenderer.pos(0.0D, f13 * d1, width).tex(minU, f11).endVertex();
-            worldrenderer.pos(d0, f13 * d1, width).tex(maxU, f11).endVertex();
-            worldrenderer.pos(d0, f13 * d1, 0.0D).tex(maxU, f11).endVertex();
-            worldrenderer.pos(0.0D, f13 * d1, 0.0D).tex(minU, f11).endVertex();
+            float f13 = f8 + 1.0F / (float)sizeY;
+            worldrenderer.pos(0.0D, (double)f13 * d1, (double)width).tex((double)minU, (double)f11).endVertex();
+            worldrenderer.pos(d0, (double)f13 * d1, (double)width).tex((double)maxU, (double)f11).endVertex();
+            worldrenderer.pos(d0, (double)f13 * d1, 0.0D).tex((double)maxU, (double)f11).endVertex();
+            worldrenderer.pos(0.0D, (double)f13 * d1, 0.0D).tex((double)minU, (double)f11).endVertex();
         }
 
         tess.draw();
@@ -144,12 +144,12 @@ public class ModelSprite
 
         for (int l = 0; l < sizeY; ++l)
         {
-            float f9 = (float)l / sizeY;
+            float f9 = (float)l / (float)sizeY;
             float f12 = minV + f1 * f9 + f3;
-            worldrenderer.pos(d0, f9 * d1, width).tex(maxU, f12).endVertex();
-            worldrenderer.pos(0.0D, f9 * d1, width).tex(minU, f12).endVertex();
-            worldrenderer.pos(0.0D, f9 * d1, 0.0D).tex(minU, f12).endVertex();
-            worldrenderer.pos(d0, f9 * d1, 0.0D).tex(maxU, f12).endVertex();
+            worldrenderer.pos(d0, (double)f9 * d1, (double)width).tex((double)maxU, (double)f12).endVertex();
+            worldrenderer.pos(0.0D, (double)f9 * d1, (double)width).tex((double)minU, (double)f12).endVertex();
+            worldrenderer.pos(0.0D, (double)f9 * d1, 0.0D).tex((double)minU, (double)f12).endVertex();
+            worldrenderer.pos(d0, (double)f9 * d1, 0.0D).tex((double)maxU, (double)f12).endVertex();
         }
 
         tess.draw();
